[reorg][cmd] Move bh to legacy
- New version will use gommon/dcli instead of cobra
