+++ _legacy/cmd/bh/ping.go
@@ -1,4 +1,4 @@
-package main
+package bh
 
 import (
 	"context"
