+++ _legacy/cmd/bh/run.go
@@ -1,4 +1,4 @@
-package main
+package bh
 
 import (
 	"context"
