+++ _legacy/cmd/bh/serve.go
@@ -1,4 +1,4 @@
-package main
+package bh
 
 import (
 	"context"
